+++ RELEASE-NOTES.md
@@ -10,6 +10,7 @@
 * #76 Series.forData(Iterable<T>) 
 * #80 CsvLoader - better way to include/exclude columns
 * #81 Implement min/max aggregation operations 
+* #82 Make Series iterable
 
 ## Release 0.6